Remove obsolete date field assignment  (#81)
* Remove obsolete date field assignment

We no longer have opensearch, so this is not needed. I have looked in
the backend and through the pipeline and can't find any instances of
it still being used. I've also left the field itself so it can still
be used if we have any more-legacy type code lying around

* bump version

* Add deprecated comment to obsolete field