Update schema to add new fields to passage object
We spotted issues inf iltering that have turned out to originate from
these being missing. Essentially the yql was skipping searching passages
because these fields where not there. So in the case of an exact match
query, we give it an impossible task of searching for text blocks, which
are passage specific AND for geographies which are document specific.

The issue seems to have been there on other queries but was masked by
the way the other queries balance other items being searched.