Test on latest vespa version
Recently work on search fixes was stalled when trying to use a feature
from a latter version then we were testing on. We run vespa tests in
four repos and are currently updating the version manually across these.

Rather than pinning a version for our tests, if we instead always test
on the latest we will also have more accurate coverage for what we deploy
into vespa cloud, as the versions there are bumped automatically.

Pinning latest is obviously bad practice for production environments
because it gives less control over what is live, however we are just
using this image for our tests, where the concern doesnt apply.