Add notify hook and iterate_batch to sdk (#197)
* Add `iterate_batch` to cpr_sdk

This function has been proving useful for quite a few of our pipelines.
Adding it to the sdk makes it more generally available across repos.

* Adding slack hook to sdk

* Fix pyright error

CI was failing on this line with the error:
```
error: Argument of type "float | None" cannot be assigned to parameter
"number" of type "_SupportsRound2[_T@round]" in function "round"
```

As the definition of `hits.relevance` says it can be None, it was causing
issues for pyright. Interestingly, only in ci though

Co-authored-by: Kalyan Dutia <kalyan.dutia@gmail.com>

* fix url mock

* remove batch_size default

---------

Co-authored-by: Kalyan Dutia <kalyan.dutia@gmail.com>