feat: Choose distance threshold that passes tests
This sets the default distance threshold to one that passes the search
tests for passage thresholds. I have made the value the highest possible
that would still pass the tests as this seems like a safer way to add
this behaviour. Later I'd like to have more test data and refine this
further and with more confidence