Feature/sci 419 no lower limit on semantic search result relevancy (#214)
This adjusts the threshold for semantic search to exclude passages in the test set that shouldn't be there, while including ones that should. For this I've built on top of the excellent data science work that was already done. Both in setting up the search intentions testing, and in making the distance threshold editable. It also relies on data identified by the policy team for an acceptable value. This is only a preliminary value to fix an issue blocking knowledge graph deployment, and I'd like to revise it further asap with more data from policy and input from data science. For this stage I've set it higher than I would otherwise having been given the steer that false positives are better than false negatives.