test: support async test cert handling
There was a previus workaround put in place for the fact that
`VespaSearchAdapter` expects certs to be passed in. This workaround fails
when running async because the certs are actually checked in this context
this moves to a simpler workaround that just makes local running simpler.
This will also have uses in other local run contexts