feat: add async vespa query
Introduces the option to run the vespa query async. For now the previous
implementation is maintained, and the async version is otehrwise exactly
the same. This allows us to try it in production for a while and see if
it addresses some of the issues we've been facing with queries.

We should come back and delete whichever one we stop using.